refactor: reorganized files structure
